Reimplement `duped`.
The old `db` feature is removed, because the new algorithm is way
faster.

The new implementation hashes files incrementally. In the first round
the first 16KiBs of all files are hashed, then in the second round the
library selects only those files that had duplicate hashes in the
first round and so on. In a few round of hashes it's going to become
obvious which files are truly duplicates, which the non-duplicate
files will be dealt with quickly.